requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
loguru==0.7.2
python-decouple==3.8
requests==2.32.3
urllib3==2.2.2
